+++ src/configure_settings/configure_settings.jl
@@ -22,7 +22,6 @@ function default_settings()
         "MultiStage" => 0,
         "MethodofMorris" => 0,
         "IncludeLossesInESR" => 0,
-        "HydrogenHourlyMatching" => 0,
         "EnableJuMPStringNames" => false,
         "HydrogenHourlyMatching" => 0,
         "DC_OPF" => 0,