+++ src/model/core/discharge/investment_discharge.jl
@@ -28,7 +28,7 @@ In addition, this function adds investment and fixed O\&M related costs related 
 \begin{aligned}
 & 	\sum_{y \in \mathcal{G} } \sum_{z \in \mathcal{Z}}
 	\left( (\pi^{INVEST}_{y,z} \times \overline{\Omega}^{size}_{y,z} \times  \Omega_{y,z})
-	+ (\pi^{FOM}_{y,z} \times  \Delta^{total}_{y,z})\right)
+	+ (\pi^{FOM}_{y,z} \times \overline{\Omega}^{size}_{y,z} \times  \Delta^{total}_{y,z})\right)
 \end{aligned}
 ```
 """
@@ -40,20 +40,11 @@ function investment_discharge!(EP::Model, inputs::Dict, setup::Dict)
 	gen = inputs["RESOURCES"]
 
 	G = inputs["G"] # Number of resources (generators, storage, DR, and DERs)
-
-	NEW_CAP = inputs["NEW_CAP"] # Set of all resources eligible for new capacity
-	RET_CAP = inputs["RET_CAP"] # Set of all resources eligible for capacity retirements
-	COMMIT = inputs["COMMIT"] # Set of all resources eligible for unit commitment
-	RETRO = inputs["RETRO"]     # Set of all retrofit resources
-
-	# Additional retrofit information if necessary
-	if !isempty(RETRO)
-		NUM_RETRO_SOURCES = inputs["NUM_RETROFIT_SOURCES"]       # The number of source resources for each retrofit resource
-		RETRO_SOURCES = inputs["RETROFIT_SOURCES"]               # Source technologies (Resource Name) for each retrofit [1:G]
-		RETRO_SOURCE_IDS = inputs["RETROFIT_SOURCE_IDS"]         # Source technologies (IDs) for each retrofit [1:G]
-		RETRO_INV_CAP_COSTS = inputs["RETROFIT_INV_CAP_COSTS"]   # The set of investment costs (capacity $/MWyr) of each retrofit by source
-		RETRO_EFFICIENCY = inputs["RETROFIT_EFFICIENCIES"]       # Ratio of installed retrofit capacity to retired source capacity [0:1]
-	end
+	
+	NEW_CAP     = inputs["NEW_CAP"] # Set of all resources eligible for new capacity
+	RET_CAP     = inputs["RET_CAP"] # Set of all resources eligible for capacity retirements
+	COMMIT      = inputs["COMMIT"] # Set of all resources eligible for unit commitment
+	RETROFIT_CAP   = inputs["RETROFIT_CAP"]  # Set of all resources being retrofitted
 
 	### Variables ###
 
@@ -67,13 +58,9 @@ function investment_discharge!(EP::Model, inputs::Dict, setup::Dict)
 		@variable(EP, vEXISTINGCAP[y=1:G] >= 0);
 	end
 
-	# Capacity from source resource "yr" that is being retrofitted into capacity of retrofit resource "r"
-	if !isempty(RETRO)
-		# Dependent iterators only allowed in forward sequence, so we reconstruct retrofit destinations from sources.
-		ALL_SOURCES = intersect(collect(Set(collect(Iterators.flatten(RETRO_SOURCE_IDS)))),RET_CAP)
-		DESTS_BY_SOURCE = [ y in ALL_SOURCES ? intersect(findall(x->in(inputs["RESOURCE_NAMES"][y],RETRO_SOURCES[x]), 1:G), findall(x->x in NEW_CAP, 1:G)) : []  for y in 1:G]
-		@variable(EP, vRETROFIT[yr in ALL_SOURCES, r in DESTS_BY_SOURCE[yr]] >= 0);     # Capacity retrofitted from source technology y to retrofit technology r
-	end
+	# Being retrofitted capacity of resource y 
+	@variable(EP, vRETROFITCAP[y in RETROFIT_CAP] >= 0);
+
 
 	### Expressions ###
 
@@ -83,54 +70,54 @@ function investment_discharge!(EP::Model, inputs::Dict, setup::Dict)
 		@expression(EP, eExistingCap[y in 1:G], existing_cap_mw(gen[y]))
 	end
 
-	# Cap_Size is set to 1 for all variables when unit UCommit == 0
-	# When UCommit > 0, Cap_Size is set to 1 for all variables except those where THERM == 1
 	@expression(EP, eTotalCap[y in 1:G],
-		if y in intersect(NEW_CAP, RET_CAP) # Resources eligible for new capacity and retirements
-			if y in COMMIT
-				eExistingCap[y] + cap_size(gen[y])*(EP[:vCAP][y] - EP[:vRETCAP][y])
-			else
-				eExistingCap[y] + EP[:vCAP][y] - EP[:vRETCAP][y]
-			end
-		elseif y in setdiff(NEW_CAP, RET_CAP) # Resources eligible for only new capacity
-			if y in COMMIT
-				eExistingCap[y] + cap_size(gen[y])*EP[:vCAP][y]
-			else
-				eExistingCap[y] + EP[:vCAP][y]
-			end
-		elseif y in setdiff(RET_CAP, NEW_CAP) # Resources eligible for only capacity retirements
-			if y in COMMIT
-				eExistingCap[y] - cap_size(gen[y])*EP[:vRETCAP][y]
-			else
-				eExistingCap[y] - EP[:vRETCAP][y]
-			end
-		else # Resources not eligible for new capacity or retirements
-			eExistingCap[y] + EP[:vZERO]
-		end
-	)
-
-	## Objective Function Expressions ##
-
-	# Fixed costs for resource "y" = annuitized investment cost plus fixed O&M costs
-	# If resource is not eligible for new capacity, fixed costs are only O&M costs
-	@expression(EP, eCFix[y in 1:G],
-		if y in setdiff(NEW_CAP, RETRO) # Resources eligible for new capacity (Non-Retrofit)
-			if y in COMMIT
-				inv_cost_per_mwyr(gen[y])*cap_size(gen[y])*vCAP[y] + fixed_om_cost_per_mwyr(gen[y])*eTotalCap[y]
-			else
-				inv_cost_per_mwyr(gen[y])*vCAP[y] + fixed_om_cost_per_mwyr(gen[y])*eTotalCap[y]
-			end
-		elseif y in intersect(NEW_CAP, RETRO) # Resources eligible for new capacity (Retrofit yr -> y)
-			if y in COMMIT
-				sum( RETRO_SOURCE_IDS[y][i] in RET_CAP ? RETRO_INV_CAP_COSTS[y][i]*cap_size(gen[y])*vRETROFIT[RETRO_SOURCE_IDS[y][i],y]*RETRO_EFFICIENCY[y][i] : 0 for i in 1:NUM_RETRO_SOURCES[y]) + fixed_om_cost_per_mwyr(gen[y])*eTotalCap[y]
-			else
-				sum( RETRO_SOURCE_IDS[y][i] in RET_CAP ? RETRO_INV_CAP_COSTS[y][i]*vRETROFIT[RETRO_SOURCE_IDS[y][i],y]*RETRO_EFFICIENCY[y][i] : 0 for i in 1:NUM_RETRO_SOURCES[y]) + fixed_om_cost_per_mwyr(gen[y])*eTotalCap[y]
-			end
+	if y in intersect(NEW_CAP, RET_CAP, RETROFIT_CAP) # Resources eligible for new capacity, retirements and being retrofitted
+		if y in COMMIT
+			eExistingCap[y] + cap_size(gen[y])*(EP[:vCAP][y] - EP[:vRETCAP][y] - EP[:vRETROFITCAP][y])
 		else
-			fixed_om_cost_per_mwyr(gen[y])*eTotalCap[y]
+			eExistingCap[y] + EP[:vCAP][y] - EP[:vRETCAP][y] - EP[:vRETROFITCAP][y]
 		end
-	)
+	elseif y in intersect(setdiff(RET_CAP, NEW_CAP), setdiff(RET_CAP, RETROFIT_CAP)) # Resources eligible for only capacity retirements
+		if y in COMMIT
+			eExistingCap[y] - cap_size(gen[y])*EP[:vRETCAP][y]
+		else
+			eExistingCap[y] - EP[:vRETCAP][y]
+		end
+	elseif y in setdiff(intersect(RET_CAP, NEW_CAP), RETROFIT_CAP) # Resources eligible for retirement and new capacity
+		if y in COMMIT
+			eExistingCap[y] + cap_size(gen[y])* (EP[:vCAP][y] - EP[:vRETCAP][y])
+		else
+			eExistingCap[y] + EP[:vCAP][y] - EP[:vRETCAP][y]
+		end
+	elseif y in setdiff(intersect(RET_CAP, RETROFIT_CAP), NEW_CAP) # Resources eligible for retirement and retrofitting
+		if y in COMMIT
+			eExistingCap[y] - cap_size(gen[y]) * (EP[:vRETROFITCAP][y] + EP[:vRETCAP][y])
+		else
+			eExistingCap[y] - (EP[:vRETROFITCAP][y] + EP[:vRETCAP][y])
+		end
+	elseif y in intersect(setdiff(NEW_CAP, RET_CAP),setdiff(NEW_CAP, RETROFIT_CAP))  # Resources eligible for only new capacity
+		if y in COMMIT
+			eExistingCap[y] + cap_size(gen[y])*EP[:vCAP][y]
+		else
+			eExistingCap[y] + EP[:vCAP][y]
+		end
+	else # Resources not eligible for new capacity or retirement
+		eExistingCap[y] + EP[:vZERO]
+	end
+)
 
+	### Need editting ##
+	@expression(EP, eCFix[y in 1:G],
+	if y in NEW_CAP # Resources eligible for new capacity (Non-Retrofit)
+		if y in COMMIT
+			inv_cost_per_mwyr(gen[y])*cap_size(gen[y])*vCAP[y] + fixed_om_cost_per_mwyr(gen[y])*eTotalCap[y]
+		else
+			inv_cost_per_mwyr(gen[y])*vCAP[y] + fixed_om_cost_per_mwyr(gen[y])*eTotalCap[y]
+		end
+	else
+		fixed_om_cost_per_mwyr(gen[y])*eTotalCap[y]
+	end
+)
 	# Sum individual resource contributions to fixed costs to get total fixed costs
 	@expression(EP, eTotalCFix, sum(EP[:eCFix][y] for y in 1:G))
 
@@ -155,6 +142,8 @@ function investment_discharge!(EP::Model, inputs::Dict, setup::Dict)
 	# Cannot retire more capacity than existing capacity
 	@constraint(EP, cMaxRetNoCommit[y in setdiff(RET_CAP,COMMIT)], vRETCAP[y] <= eExistingCap[y])
 	@constraint(EP, cMaxRetCommit[y in intersect(RET_CAP,COMMIT)], cap_size(gen[y])*vRETCAP[y] <= eExistingCap[y])
+	@constraint(EP, cMaxRetroNoCommit[y in setdiff(RETROFIT_CAP,COMMIT)], vRETROFITCAP[y] + vRETCAP[y] <= eExistingCap[y])
+	@constraint(EP, cMaxRetroCommit[y in intersect(RETROFIT_CAP,COMMIT)], cap_size(gen[y]) * (vRETROFITCAP[y] + vRETCAP[y]) <= eExistingCap[y])
 
 	## Constraints on new built capacity
 	# Constraint on maximum capacity (if applicable) [set input to -1 if no constraint on maximum capacity]